Added spritesheet util and two basic sprite sheets
- started work on the constant points for sprites to make it easier to
find specific sprites. (i.e. grass sprite point, rock sprite point) can
find the beginnings of it in the spriteConstant subfolder to the util
page.
- Added some material type enums that will be used later for the
inventory and item system.
- Began work on the designer panel that will go on the other side of the
main window and will have a sroll bar with lists of terrains or objects
that can be selected and applied to the world through the viewer window.